231114 회원정보 보기, 수정, 삭제시 권한 확인

+++ src/main/java/com/example/prj1be231109/controller/MemberController.java
@@ -64,16 +64,33 @@ public class MemberController {
     }
     
     @GetMapping("")
-    public ResponseEntity<Member> view(String id) {
+    public ResponseEntity<Member> view(String id,
+                                       @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         Member member = service.getMember(id);
 
         return  ResponseEntity.ok(member);
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
         // todo : 로그인 했는지? -> 안했으면 401
         // todo : 자기 정보인지? -> 아니면 403
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -83,8 +100,16 @@ public class MemberController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
+    public ResponseEntity edit(@RequestBody Member member,
+                               @SessionAttribute(value = "login", required = false) Member login) {
         // TODO: 로그인 했는지" 자기 정보인지?
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(member.getId(), login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         if (service.update(member)) {
             return ResponseEntity.ok().build();
